Migrated Character Save/load methods to Ch. Class
