+++ Character.cs
@@ -3,11 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Threading; // timeout
+using System.Text.Json; // has to be installed in nuget Package
+using System.IO;  // to create and read files
 
 namespace RpgGame
 {
-  internal class Character
-  {
+  internal class Character :RegexMethods {
     // Klassenvariabeln
 
 
@@ -82,6 +84,185 @@ namespace RpgGame
     public uint[] Exp { get; set; }
 
     public byte Lvl { get; set; }
+
+    /// <summary>
+    /// Creates an new Character with Name and Class.<br />
+    /// </summary>
+    /// <returns>Character</returns>
+    public static Character CreateCharacter() {
+      string name = "";
+      byte cl = 0;
+
+      while (name == "") {
+        Console.Clear();
+        Console.WriteLine("Geben Sie den Namen ihres Charakters ein:");
+        name = Console.ReadLine();
+
+        if (IsInValidSign(name)) {
+          Console.WriteLine("\nIm Namen ist ein unerlaubtes Zeichen enthalten!");
+          Thread.Sleep(500);
+          continue;
+        } else if (name == "" || name == " ") {
+          Console.WriteLine("\nDer Name darf nicht leer sein!");
+          Thread.Sleep(500);
+          continue;
+        }
+
+        // convert to char array of the string
+        char[] letters = name.ToCharArray();
+        // upper case the first char
+        letters[0] = char.ToUpper(letters[0]);
+        // put array back together
+        name = new string(letters);
+      }
+
+      do {
+        Console.Clear();
+        Console.WriteLine("Was ist die Klasse ihres Charakters?\n1) Krieger\n2) Magier\n3) Schurke");
+        cl = Convert.ToByte(Console.ReadKey(false).KeyChar - 48);
+      } while (cl < 1 || cl > 4);
+
+      return new Character(name, cl);
+    }
+
+    /// <summary>
+    /// Saves Json File with current Character Stats<br />
+    /// https://www.nuget.org/packages/System.Text.Json<br />
+    /// </summary>
+    /// <param name="c">current character</param>
+    public static void SaveCharacter(Character c) {
+      Character _characterData = new Character(c.Name, 0) {
+        Name = c.Name,
+        Class = c.Class,
+        Strength = c.Strength,
+        Intelligents = c.Intelligents,
+        Dexterity = c.Dexterity,
+        CritChance = c.CritChance,
+        CritDmg = c.CritDmg,
+        Health = c.Health,
+        Gold = c.Gold,
+        Exp = c.Exp,
+        Lvl = c.Lvl,
+      };
+
+      string path = Directory.GetCurrentDirectory();  // current Path
+      string json = JsonSerializer.Serialize(_characterData);
+
+      try {
+        File.WriteAllText(path + @"\character_saves\" + c.Name + @".json", json);
+        Console.Clear();
+        Console.WriteLine("Speichern erfolgreich.");
+      } catch (InvalidCastException e) { }
+
+      Thread.Sleep(600);
+    }
+
+    public static bool HasCharacters() {
+      // https://www.geeksforgeeks.org/c-sharp-program-for-listing-the-files-in-a-directory/
+      string path = Directory.GetCurrentDirectory() + @"\character_saves\";  // current Path
+      DirectoryInfo characterSaves = new DirectoryInfo(path);
+      FileInfo[] Files = characterSaves.GetFiles();
+
+      if (Files == null) return false;
+      return true;
+    }
+
+    public static Character GetCharacters() {
+      // https://www.geeksforgeeks.org/c-sharp-program-for-listing-the-files-in-a-directory/
+      string path = Directory.GetCurrentDirectory() + @"\character_saves\";  // current Path
+      DirectoryInfo characterSaves = new DirectoryInfo(path);
+      FileInfo[] Files = characterSaves.GetFiles();
+      List<Character> charactersList = new List<Character>();
+      byte choosenCharacterId = 0;
+
+      // fill character list
+      // https://www.tutorialsteacher.com/articles/convert-json-string-to-object-in-csharp
+      foreach (FileInfo i in Files) {
+        string jsonCharacterData = File.ReadAllText(path + i);
+        charactersList.Add(JsonSerializer.Deserialize<Character>(jsonCharacterData));
+      }
+
+      // list all characters
+      Character[] characters = charactersList.ToArray();  // convert list to array
+      ListCharacters(characters);
+
+      choosenCharacterId = ChooseCharacter(); // player input
+
+      if (choosenCharacterId == 0) return CreateCharacter();  // create new character
+
+      // decrease id by one to be sync with the array
+      if (CanLoadCharacter(--choosenCharacterId, characters)) return LoadCharacter(--choosenCharacterId, characters);
+      else throw new IndexOutOfRangeException("Die geladene Characterdatei ist korrput.");
+    }
+    private static void ListCharacters(Character[] characters) {
+      Console.WriteLine("Welcher Charakter soll geladen werden:");
+      Console.WriteLine("0) keiner (neuen Character erstellen)");
+
+      for (byte i = 0; i < characters.Length; i++) {
+        if (i == 255) break;
+        Console.WriteLine("{0}) {1}, {2} (Level: {3})",
+          i + 1, characters[i].Name, characters[i].GetClassName(), characters[i].Lvl
+        );
+      }
+    }
+
+    private static byte ChooseCharacter() {
+      byte input = 0;
+
+      do {
+        Console.Write("Eingabe: ");
+      } while (!byte.TryParse(Console.ReadLine(), out input));
+
+      return input;
+    }
+
+    /// <summary>
+    /// checks if charactersave if correct and can be loaded
+    /// </summary>
+    /// <param name="characterId">Id of loading Character</param>
+    /// <param name="characters">Array of all Characters</param>
+    /// <returns></returns>
+    private static bool CanLoadCharacter(byte characterId, Character[] characters) {
+      if (IsCharacterValid(characters[characterId])) return true;
+
+      return false;
+    }
+
+    /// <summary>
+    /// Checks if the loaded save wasnt modified.<br />
+    /// If it was modified, it cannot be loaded
+    /// </summary>
+    /// <param name="c">characer object</param>
+    /// <returns>true - if character is valid / flase - if not</returns>
+    private static bool IsCharacterValid(Character c) {
+      bool nameVaild = false, classValid = false;
+
+      if (c.Name == "" || IsInValidSign(c.Name)) nameVaild = true;
+      if (c.Class > 1 && c.Class < 4) classValid = true;
+      c.Strength = c.Strength;
+      c.Intelligents = c.Intelligents;
+      c.Dexterity = c.Dexterity;
+      c.CritChance = c.CritChance;
+      c.CritDmg = c.CritDmg;
+      c.Health = c.Health;
+      c.Gold = c.Gold;
+      c.Exp = c.Exp;
+      c.Lvl = c.Lvl;
+
+      if (nameVaild && classValid) return true;
+
+      return false;
+    }
+
+    /// <summary>
+    /// Return the character
+    /// </summary>
+    /// <param name="characterId">Id of loading Character</param>
+    /// <param name="characters">Array of all Characters</param>
+    /// <returns></returns>
+    private static Character LoadCharacter(byte characterId, Character[] characters) {
+      return characters[characterId];
+    }
 
     /// <summary>
     /// Returns Classname
